requirements.txt
flask-sqlalchemy
psycopg2
flask-script
flask-migrate
Flask-Failsafe
flask-jwt
flask-restful
Flask-Mail
Flask-Cors
